Move `display_*`-functions to own module
